+++ components/core/FontFrame/index.js
@@ -3,8 +3,9 @@ import * as React from "react";
 import { css } from "@emotion/react";
 
 import { useFont, useFontControls } from "./hooks";
-import { Controls } from "./Controls/index";
-import { FixedControls, MobileFixedControls } from "./Controls/FixedControls";
+import { Controls } from "./Settings/index";
+import { FixedControls, MobileFixedControls } from "./Settings/FixedControls";
+import FontView from "./Views/index";
 
 const GET_STYLES_CONTAINER = (theme) => css`
   position: relative;
@@ -15,6 +16,7 @@ const GET_STYLES_CONTAINER = (theme) => css`
   background-color: ${theme.fontPreviewDarkMode ? theme.system.pitchBlack : theme.system.white};
   padding-top: 14px;
 `;
+
 const STYLES_MOBILE_HIDDEN = (theme) => css`
   @media (max-width: ${theme.sizes.mobile}px) {
     display: none;
@@ -27,6 +29,21 @@ const STYLES_MOBILE_ONLY = (theme) => css`
   }
 `;
 
+const STYLES_FONT_LOADER = (theme) => css`
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  position: absolute;
+  width: 100%;
+  height: 100%;
+  color: ${theme.fontPreviewDarkMode ? "#fff" : "#000"};
+  background-color: ${theme.fontPreviewDarkMode ? "rgba(0,0,0,0.5)" : "rgba(255,255,255,0.8)"};
+`;
+const FontLoader = () => (
+  <div css={STYLES_FONT_LOADER}>
+    <p>loading...</p>
+  </div>
+);
 export default function FontFrame({ cid, url, ...props }) {
   const { isFontLoading, fontName } = useFont({ url, name: cid }, [cid]);
   const [
@@ -42,6 +59,7 @@ export default function FontFrame({ cid, url, ...props }) {
       updateColumn,
       updateTextAlign,
       updateVerticalAlign,
+      updateCustomView,
       resetLayout,
       getRandomLayout,
     },
@@ -59,51 +77,14 @@ export default function FontFrame({ cid, url, ...props }) {
         />
       </div>
       <div style={{ position: "relative", flexGrow: 1, overflowY: "scroll" }}>
-        {isFontLoading && (
-          <div
-            css={(theme) =>
-              css({
-                display: "flex",
-                justifyContent: "center",
-                alignItems: "center",
-                position: "absolute",
-                width: "100%",
-                height: "100%",
-                color: theme.fontPreviewDarkMode ? "#fff" : "#000",
-                backgroundColor: theme.fontPreviewDarkMode
-                  ? "rgba(0,0,0,0.5)"
-                  : "rgba(255,255,255,0.8)",
-              })
-            }
-          >
-            <p>loading...</p>
-          </div>
-        )}
-        {currentState.view === "glyphs" && <Glyphs />}
-        {currentState.view === "sentence" && (
-          <Sentence
-            content="The sun is gone but I have a light The day is done but I'm having fun I think I'm dumb or maybe I'm just happy I think I'm just happy I think I'm just happy I'm not like them but I can pretend The sun is gone but I have a light The day is done but I'm having fun I think I'm dumb or maybe I'm just happy"
-            valign={currentState.context.settings.valign}
-            textAlign={currentState.context.settings.textAlign}
-            fontSize={currentState.context.settings.fontSize}
-            lineHeight={currentState.context.settings.lineHeight}
-            tracking={currentState.context.settings.tracking}
-          />
-        )}
-        {currentState.view === "paragraph" && (
-          <Paragraph
-            content="The sun is gone but I have a light The day is done but I'm having fun I think I'm dumb or maybe I'm just happy I think I'm just happy I think I'm just happy I'm not like them but I can pretend The sun is gone but I have a light The day is done but I'm having fun I think I'm dumb or maybe I'm just happy I think I'm just happy I think I'm just happy I think I'm just happy My heart is broke but I have some glue Help me inhale and mend it with you We'll float around and hang out on clouds Then we'll come down and have a hangover We'll have a hangover We'll have a hangover We'll have a hangover Skin the sun, fall asleep
-            The sun is gone but I have a light The day is done but I'm having fun I think I'm dumb or maybe I'm just happy I think I'm just happy I think I'm just happy I'm not like them but I can pretend The sun is gone but I have a light The day is done but I'm having fun I think I'm dumb or maybe I'm just happy I think I'm just happy I think I'm just happy I think I'm just happy My heart is broke but I have some glue Help me inhale and mend it with you We'll float around and hang out on clouds Then we'll come down and have a hangover We'll have a hangover We'll have a hangover We'll have a hangover Skin the sun, fall asleep
-            The sun is gone but I have a light The day is done but I'm having fun I think I'm dumb or maybe I'm just happy I think I'm just happy I think I'm just happy I'm not like them but I can pretend The sun is gone but I have a light The day is done but I'm having fun I think I'm dumb or maybe I'm just happy I think I'm just happy I think I'm just happy I think I'm just happy My heart is broke but I have some glue Help me inhale and mend it with you We'll float around and hang out on clouds Then we'll come down and have a hangover We'll have a hangover We'll have a hangover We'll have a hangover Skin the sun, fall asleep
-            "
-            valign={currentState.context.settings.valign}
-            textAlign={currentState.context.settings.textAlign}
-            fontSize={currentState.context.settings.fontSize}
-            lineHeight={currentState.context.settings.lineHeight}
-            tracking={currentState.context.settings.tracking}
-            column={currentState.context.settings.column}
-          />
-        )}
+        {isFontLoading && <FontLoader />}
+        <FontView
+          view={currentState.view}
+          customView={currentState.customView}
+          customViewContent={currentState.context.customViewContent}
+          settings={currentState.context.settings}
+          updateCustomView={updateCustomView}
+        />
       </div>
       <div css={STYLES_MOBILE_HIDDEN}>
         {currentState.context.showSettings && (
@@ -135,110 +116,3 @@ export default function FontFrame({ cid, url, ...props }) {
     </div>
   );
 }
-
-const STYLES_SENTENCE = (theme) => css`
-  width: 100%;
-  margin-top: 12px;
-  color: ${theme.fontPreviewDarkMode ? theme.system.white : theme.system.pitchBlack};
-  padding: 0px 32px 28px;
-  &:focus {
-    outline: none;
-  }
-`;
-const Sentence = ({ content, valign, textAlign, fontSize, lineHeight, tracking }) => {
-  const [value, setValue] = React.useState(`<p>${content}</p>`);
-  const mapAlignToFlex = { center: "center", top: "flex-start", bottom: "flex-end" };
-  return (
-    <div style={{ display: "flex", alignItems: mapAlignToFlex[valign], height: "100%" }}>
-      <div
-        contentEditable="true"
-        suppressContentEditableWarning={true}
-        style={{
-          fontSize: `${fontSize}px`,
-          lineHeight: `${lineHeight}%`,
-          letterSpacing: `${tracking}em`,
-          textAlign,
-        }}
-        css={STYLES_SENTENCE}
-        onKeyDown={(e) => {
-          e.stopPropagation();
-        }}
-      >
-        {content}
-      </div>
-    </div>
-  );
-};
-
-const Paragraph = ({ content, valign, textAlign, fontSize, lineHeight, tracking, column }) => {
-  const [value, setValue] = React.useState(`<p>${content}</p>`);
-  const mapAlignToFlex = { center: "center", top: "flex-start", bottom: "flex-end" };
-  return (
-    <div style={{ display: "flex", alignItems: mapAlignToFlex[valign], height: "100%" }}>
-      <div
-        contentEditable="true"
-        suppressContentEditableWarning={true}
-        style={{
-          fontSize: `${fontSize}px`,
-          lineHeight: `${lineHeight}%`,
-          letterSpacing: `${tracking}em`,
-          textAlign,
-          whiteSpace: "pre-wrap",
-        }}
-        css={[
-          STYLES_SENTENCE,
-          css`
-            width: 100%;
-            column-count: ${column};
-            column-gap: 24px;
-          `,
-        ]}
-        onKeyDown={(e) => {
-          e.stopPropagation();
-        }}
-      >
-        {content}
-      </div>
-    </div>
-  );
-};
-
-const STYLES_GLYPHS_WRAPPER = (theme) => css`
-  padding: 0px 32px 28px;
-  color: ${theme.fontPreviewDarkMode ? theme.system.white : theme.system.pitchBlack};
-`;
-const STYLES_GLYPHS_LETTER = css`
-  font-size: 128px;
-  line-height: 192px;
-  margin-bottom: 16px;
-`;
-const STYLES_GLYPHS_GRID = css`
-  margin-top: -16px;
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(12px, 1fr));
-  grid-template-rows: repeat(12, 1fr);
-  grid-column-gap: 28px;
-  grid-auto-rows: 0px;
-  overflow: hidden;
-`;
-const Glyphs = ({}) => {
-  const content = `ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789!?()@$#%*[]{}\:;_""-`;
-  const glyphs = React.useMemo(() => new Array(6).fill(content).join("").split(""), []);
-  return (
-    <div css={STYLES_GLYPHS_WRAPPER}>
-      <div css={STYLES_GLYPHS_LETTER}>Aa</div>
-      <div css={STYLES_GLYPHS_GRID}>
-        {glyphs.map((letter, i) => (
-          <div
-            key={letter + i}
-            css={css`
-              margin-top: 16px;
-            `}
-          >
-            {letter}
-          </div>
-        ))}
-      </div>
-    </div>
-  );
-};
